conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bcrypt=4.1.2=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
bottleneck=1.3.5=py310h9128911_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2024.2.2=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cryptography=41.0.3=py310h3438e0d_0
cuda-cccl=12.2.140=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.2.140=0
cuda-runtime=11.8.0=0
cudatoolkit=11.8.0=hd77b12b_0
cycler=0.12.1=pypi_0
cython=3.0.8=pypi_0
dataclasses-json=0.6.3=pypi_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
distro=1.9.0=pypi_0
dnspython=2.4.2=pypi_0
eigen=3.4.0=h2d74725_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
extra-streamlit-components=0.1.60=pypi_0
ffmpeg=4.2.3=ha925a31_0
filelock=3.9.0=py310haa95532_0
fonttools=4.45.1=pypi_0
freetype=2.10.4=h546665d_1
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h8d14728_2
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
glib=2.69.1=h5dc1a3c_2
grad-cam=1.5.0=pypi_0
greenlet=3.0.3=pypi_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h11=0.14.0=pypi_0
hdf5=1.12.1=h51c971a_3
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.2=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
imageio=2.31.3=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.18.1=pyh5737063_1
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h8ffe710_2
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py310h5588dad_0
kiwisolver=1.4.5=pypi_0
langchain=0.1.1=pypi_0
langchain-community=0.0.13=pypi_0
langchain-core=0.1.11=pypi_0
langchain-openai=0.0.2.post1=pypi_0
langsmith=0.0.81=pypi_0
lazy-loader=0.3=pypi_0
lerc=3.0=hd77b12b_0
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.3.141=0
libcurand-dev=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.17=h8ffe710_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py310h2bbff1b_0
marshmallow=3.20.2=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpmath=1.3.0=py310haa95532_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=py310haa95532_0
numexpr=2.8.7=py310h2cd9be0_0
numpy=1.26.0=pypi_0
numpy-base=1.25.2=py310h65a83cf_0
openai=1.8.0=pypi_0
opencv=4.6.0=py310ha36de5b_5
opencv-python=4.9.0.80=pypi_0
openssl=1.1.1w=h2bbff1b_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py310h4ed8f06_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=pypi_0
pip=23.2.1=py310haa95532_0
platformdirs=4.0.0=pyhd8ed1ab_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=4.25.2=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.2=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydeck=0.8.1b0=pypi_0
pydensecrf=1.0rc3=py310hf5e1058_3
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pymongo=4.6.0=pypi_0
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.0=h96c0403_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=2.0.1=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py310haa95532_0
pywavelets=1.4.1=pypi_0
pywin32=305=py310h2bbff1b_0
pyyaml=6.0.1=pypi_0
pyzmq=23.2.1=py310h73ada01_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
referencing=0.32.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=py310haa95532_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
safetensors=0.4.1=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=py310h4ed8f06_0
scipy=1.11.2=pypi_0
setuptools=68.0.0=py310haa95532_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post11=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.30.0=pypi_0
streamlit-authenticator=0.2.3=pypi_0
streamlit-image-comparison=0.0.4=pypi_0
sympy=1.11.1=py310haa95532_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.9.18=pypi_0
tiktoken=0.5.2=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.15.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchaudio=2.0.2=pypi_0
torchray=1.0.0.2=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.2=py310he2412df_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.14.0=pyhd8ed1ab_0
transformers=4.36.2=pypi_0
ttach=0.0.3=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.7.1=py310haa95532_0
tzdata=2023c=h04d1e81_0
tzlocal=5.2=pypi_0
urllib3=1.26.16=py310haa95532_0
validators=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.4=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
